requirements.txt
Django==5.1.3
django-dotenv==1.4.2
django-environ==0.11.2
django-tailwind==3.8.0
djangorestframework==3.15.2
drf-yasg==1.21.8
gunicorn==23.0.0
psutil==6.1.0
psycopg2-binary==2.9.10
python-decouple==3.8
django-cors-headers==4.6.0
